fix!: update attribute type in AFD
BREAKING CHANGE: AFD.codINEP is now an Integer (previous long).

+++ src/main/java/edu/apigestor/entity/domain/AFD.java
@@ -20,7 +20,7 @@ public class AFD {
 
   @Id
   @Column(name = "codEscola")
-  private long codINEP;
+  private int codINEP;
 
   @Column(name = "tipo")
   private int tipo;
